move localizedDiscount to VariableDataProvider
